ai-agent-platform/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.25.2
beautifulsoup4==4.12.2
stripe==7.4.0
pytest==7.4.3
pytest-asyncio==0.21.1
anthropic==0.7.8
google-generativeai==0.3.2
playwright==1.40.0
scrapy==2.11.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
jobspy==2.4.0
browser-use==0.1.1
pandas==2.2.0
langchain-google-genai==1.0.3